requirements.txt
faster_whisper
python-telegram-bot
pyyaml
